add component player-list working + add room controller and config file + update css organisation

+++ TheBlindTest/src/main/java/com/theblindtest/demo/services/RoomService.java
@@ -1,5 +1,6 @@
 package com.theblindtest.demo.services;
 
+import com.theblindtest.demo.entities.Player;
 import com.theblindtest.demo.entities.Room;
 import com.theblindtest.demo.repositories.RoomRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,10 +18,6 @@ public class RoomService {
         return roomRepository.save(room);
     }
 
-    public List<Room> saveRooms(List<Room> players) {
-        return roomRepository.saveAll(players);
-    }
-
     public List<Room> getRooms() {
         return roomRepository.findAll();
     }
@@ -29,9 +26,33 @@ public class RoomService {
         return roomRepository.findById(id).orElse(null);
     }
 
+    public Room updateRoom(Room room) {
+
+        boolean isRoomExist = roomRepository.existsById(room.getId());
+
+        if (!isRoomExist) {
+            throw new IllegalStateException("la room avec l'id " + room.getId() + " n'existe pas.");
+        }
+
+        Room existingRoom = roomRepository.findById(room.getId()).orElse(null);
+
+        existingRoom.setCode(room.getCode());
+        existingRoom.setPlayers(room.getPlayers());
+
+        return roomRepository.save(existingRoom);
+    }
+
     public String deleteRoom(Long id) {
+
+        boolean isRoomExist = roomRepository.existsById(id);
+
+        if (!isRoomExist) {
+            throw new IllegalStateException("la room avec l'id " + id + " n'existe pas.");
+        }
+
         roomRepository.deleteById(id);
-        return "room supprimé ! " + id;
+
+        return "player " + id + " supprimé !";
     }
 
 }
